+++ backend/openedx_ai_extensions/processors/__init__.py
@@ -2,8 +2,8 @@
 Processors module - handles data extraction and AI processing
 """
 
-from .educator_assistant_processor import EducatorAssistantProcessor
-from .llm_processor import LLMProcessor
+from .llm.educator_assistant_processor import EducatorAssistantProcessor
+from .llm.llm_processor import LLMProcessor
 from .openedx.content_libraries_processor import ContentLibraryProcessor
 from .openedx.openedx_processor import OpenEdXProcessor
 from .openedx.submission_processor import SubmissionProcessor
