+++ backend/openedx_ai_extensions/processors/llm/llm_processor.py
@@ -8,8 +8,8 @@ import logging
 from litellm import completion, responses
 
 from openedx_ai_extensions.functions.decorators import AVAILABLE_TOOLS
-from openedx_ai_extensions.processors.litellm_base_processor import LitellmProcessor
-from openedx_ai_extensions.processors.llm_providers import adapt_to_provider, after_tool_call_adaptations
+from openedx_ai_extensions.processors.llm.litellm_base_processor import LitellmProcessor
+from openedx_ai_extensions.processors.llm.providers import adapt_to_provider, after_tool_call_adaptations
 
 logger = logging.getLogger(__name__)
