+++ backend/tests/test_component_extractors.py
@@ -110,7 +110,10 @@ def test_html_to_text_with_embeds():
 
 def test_html_to_text_invalid_html():
     """Test html_to_text returns raw HTML if parsing fails."""
-    with patch("openedx_ai_extensions.processors.component_extractors.BeautifulSoup", side_effect=Exception):
+    with patch(
+        "openedx_ai_extensions.processors.openedx.utils.component_extractors.BeautifulSoup",
+        side_effect=Exception,
+    ):
         assert html_to_text("<bad>") == "<bad>"
 
 
@@ -530,7 +533,7 @@ def test_process_problem_html_empty_input():
     assert _process_problem_html(None, False) == ""
 
 
-@patch("openedx_ai_extensions.processors.component_extractors.BeautifulSoup")
+@patch("openedx_ai_extensions.processors.openedx.utils.component_extractors.BeautifulSoup")
 def test_process_problem_html_exception_handling(mock_bs):
     """Test that if processing fails, raw HTML is returned."""
     # Force an exception during processing
